refactor: change talk quiz api controller
